Fix API tokens problems
The previous identifiers created were of type "web application", which do not give you access to a refresh token even if you ask for it apparently. As such, had to create another kind of id, with a loopback redirect url instead.

Also added a nice little animation on "Synchronize" click